+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 * Made all service names public
 * Added javadoc to all services
 * Deprecated ambiguous varargs sendRpc() methods, and replaced them with explicit versions
+* Invoke callback throws IllegalStateException if wait times out but no value has been set
+* Fixed bug where LCDS Heartbeats disconnected the client
 
 ### 1.5.5
 ##### Xmpp